backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cloudinary==1.36.0
crispy-bootstrap4==2023.1
cryptography==41.0.4
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==5.0.2
django-allauth==0.57.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-summernote==0.8.20.0
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.4
oauthlib==3.2.2
openai==1.12.0
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.0.6
webencodings==0.5.1
whitenoise==6.6.0
